OVS-95 vdisk size changed back to int - plus better variable name in processor.py +OVS-64

+++ ovs/dal/hybrids/vdisk.py
@@ -18,7 +18,7 @@ class VDisk(DataObject):
     # pylint: disable=line-too-long
     _blueprint = {'name':              (None,   str,  'Name of the virtual disk'),
                   'description':       (None,   str,  'Description of the virtual disk'),
-                  'size':              (0,      long,  'Size of the virtual disk'),
+                  'size':              (0,      int,  'Size of the virtual disk'),
                   'devicename':        (None,   str,  'The name of the container file backing the vDisk'),
                   'order':             (None,   int,  'Order of the virtual disk in which they are attached'),
                   'volumeid':          (None,   str,  'Volume ID representing the virtual disk'),
